# #12108 Speed up the twisted.web HTTP client