--- a/src/twisted/newsfragments/12112.bugfix
+++ b/src/twisted/newsfragments/12112.bugfix
@@ -1,0 +1,1 @@
+Fix bug where ``twisted.python.failure.Failure``'s custom pickling code wasn't being used.