--- a/src/twisted/python/failure.py
+++ b/src/twisted/python/failure.py
@@ -250,7 +250,7 @@ class Failure(BaseException):
 
     pickled = 0
     stack = None
-    __parents = None
+    _parents = None
 
     # The opcode of "yield" in Python bytecode. We need this in
     # _findFailure in order to identify whether an exception was
@@ -419,15 +419,15 @@ class Failure(BaseException):
 
     @property
     def parents(self):
-        if self.__parents is not None:
-            return self.__parents
+        if self._parents is not None:
+            return self._parents
 
         if inspect.isclass(self.type) and issubclass(self.type, Exception):
             parentCs = getmro(self.type)
-            self.__parents = list(map(reflect.qual, parentCs))
+            self._parents = list(map(reflect.qual, parentCs))
         else:
-            self.__parents = [self.type]
-        return self.__parents
+            self._parents = [self.type]
+        return self._parents
 
     def _extrapolate(self, otherFailure):
         """
@@ -613,11 +613,25 @@ class Failure(BaseException):
     def __str__(self) -> str:
         return "[Failure instance: %s]" % self.getBriefTraceback()
 
+    def __setstate__(self, state):
+        state["_parents"] = state.pop("parents")
+        self.__dict__.update(state)
+
     def __getstate__(self):
-        """Avoid pickling objects in the traceback."""
-        if self.pickled:
-            return self.__dict__
+        """
+        Avoid pickling objects in the traceback.
+
+        This is not called direclty by pickle, since C{BaseException}
+        implements reduce; instead, pickle calls C{Failure.__reduce__} which
+        then calls this API.
+        """
+        # Make sure _parents field is populated:
+        _ = self.parents
+
         c = self.__dict__.copy()
+
+        # Backwards compatibility with old code, e.g. for Perspective Broker:
+        c["parents"] = c.pop("_parents")
 
         c["frames"] = [
             [
@@ -649,6 +663,11 @@ class Failure(BaseException):
 
         c["pickled"] = 1
         return c
+
+    def __reduce__(self):
+        # BaseException implements a __reduce__ (in C, technically), so TODO
+        from functools import partial
+        return (partial(Failure.__new__, Failure), (), self.__getstate__())
 
     def cleanFailure(self):
         """