--- a/src/twisted/web/_newclient.py
+++ b/src/twisted/web/_newclient.py
@@ -1051,6 +1051,9 @@ def makeStatefulDispatcher(name, template):
 # IClientRequest.
 _ClientRequestProxy = proxyForInterface(IClientRequest)
 
+# Pre-create this since it gets created for every response.
+_RESPONSE_DONE_FAILURE = Failure(ResponseDone("Response body fully received"))
+
 
 @implementer(IResponse)
 class Response:
@@ -1270,7 +1273,7 @@ class Response:
         """
         self._state = "DEFERRED_CLOSE"
         if reason is None:
-            reason = Failure(ResponseDone("Response body fully received"))
+            reason = _RESPONSE_DONE_FAILURE
         self._reason = reason
 
     def _bodyDataFinished_CONNECTED(self, reason=None):
@@ -1278,7 +1281,7 @@ class Response:
         Disconnect the protocol and move to the C{'FINISHED'} state.
         """
         if reason is None:
-            reason = Failure(ResponseDone("Response body fully received"))
+            reason = _RESPONSE_DONE_FAILURE
         self._bodyProtocol.connectionLost(reason)
         self._bodyProtocol = None
         self._state = "FINISHED"