--- a/src/twisted/web/newsfragments/12108.feature
+++ b/src/twisted/web/newsfragments/12108.feature
@@ -1,0 +1,1 @@
+The HTTP client now runs faster by using a little less CPU.