@itamarst 2024-03-06
It's not clear to me there _is_ a traceback in this case, since it's directly creating exceptions. Will validate tomorrow.