@adiroiban 2024-03-07
> It's not clear to me there is a traceback in this case, since it's directly creating exceptions. Will validate tomorrow.

You are right.
There is no traceback here, so traceback is not an issue.

Still, maybe it's a good idea to have different text messages, depending on the place from where this failure is returned.