@itamarst 2024-03-11
So, the main issue is mutable `Failure`. `FrozenFailure` is certainly possible, but... it'd have to make the exception frozen somehow, when it's not clear how. So a generic version is hard. And any mutable exception attributes can be mutated even if `__setattr__` is disabled...