@itamarst 2024-03-11
OK, added a Failure subclass.