@glyph 2024-03-11
> OK, added a Failure subclass.

This is *probably* fine, but thinking about it does send a chill down my spine.

I don't think that an immutable failure *subclass* is actually substitutable, or thereby compliant with the letter of the compatibility policy, as it stands. I'd kinda rather make *all* Failures be immutable if that is practical, and initially enforce it at type-time, which is not covered by the compatibility policy?