@exarkun 2024-03-12
> > OK, added a Failure subclass.
> 
> This is _probably_ fine, but thinking about it does send a chill down my spine.
> 

*cough* CopyableFailure *cough*

> I don't think that an immutable failure subclass is actually substitutable, or thereby compliant with the letter of the compatibility policy, as it stands. I'd kinda rather make all Failures be immutable if that is practical, and initially enforce it at type-time, which is not covered by the compatibility policy?

It's certainly not.  I think _only_ enforcing it at type-time is probably good enough.  There will never be a 100% immutable Python object.  If the type checker tells someone they're not allowed to change something and they go ahead and run code that changes it anyway, that seems like a choice that's up to them.

Ideally, there'd be a generic like `typing.Readonly` - as in, `Readonly[Failure]`.  I don't think there is, but a `ReadonlyFailure` protocol that exposes all of `Failure`'s public attributes, but only for reading is still simple, just more typing.  Then I believe the definition of `freeze` is:

```
    def freeze(self) -> ReadonlyFailure: return self
```