@exarkun 2024-03-12
(This PR also seems like a weird place to be solving this 24 year old design mistake)