@itamarst 2024-03-12
https://github.com/python/typing/discussions/1123 has an example of how to define a Protocol with read-only attributes.