@itamarst 2024-03-12
Given the Failure object is going to end up in ... places, e.g. `BodyProtocol.connectionLost`, using typing to enforce read-only-ness might end up going pretty far in terms of changing the API.