@glyph 2024-03-12
> _cough_ CopyableFailure _cough_

Okay, perhaps a comparable amount of chills, but it *is* a proper Liskov substitute ;)