@glyph 2024-03-12
> Then I believe the definition of `freeze` is:

My suggestion would be to just have `Failure` itself be frozen by default, because:

1. Mutating a failure is a pretty weird operation that I would guess rarely happens?
2. Any code doing the mutating would continue to work at runtime, unless it mutates one of these shared instances (but what mutation would such code be making, I wonder?) in which case behavior would change, but not in a technically incompatible way.
3. Any code doing the mutating would become a type error, which seems like a reasonable way for users to catch it and fix things.

If my first assumption is wrong (this would not affect much code) then we should probably do another approach. Although then I'd really want to know what the common failure-mutation patterns are.