@itamarst 2024-03-12
Based on above discussion, before proceeding on merging this PR I'm going to:

1. Open a separate issue about make `Failure` (and relevant exceptions) immutable.
2. Make a sketch on a new branch.
3. Maybe even get it to review stage, we'll see.

If it still seems plausible then we'll see what order of merging separate PRs makes sense.