@itamarst 2024-03-12
I opened #12111 for the immutable `Failure` case, would appreciate if people on this thread check it to see if I missed any important comments from the discussion here.