@itamarst 2024-03-13
I gave up on reusing `Failure` instances and went with a different approach.