@glyph 2024-03-14
> > MyPy failures don't appear to have anything to do with this branch...
> 
> Confirmed, fresh trunk checkout gives similar errors here. Investigating…

filed https://github.com/twisted/twisted/issues/12113 to track and discuss