@glyph 2024-03-14
@itamarst an attempted fix is here https://github.com/twisted/twisted/pull/12115